Fetch DEM tiles as needed, over the system's HTTPS, and hold them to the survey
platform/http gets through WinHTTP, NSURLSession, or the system's libcurl
loaded at run time. world/download keeps DEM tiles in the cache directory,
checked against their ETag MD5, and the geoid pinned by SHA-256. The CLI's
height command gives the ground's height anywhere, and every package runs it
at Sydney airport. Twelve surveyed runway ends and five coastal waters are
within the DEM's stated 4 m; summits are 8 to 35 m low, as a 30 m grid is,
and that is pinned and documented. CI registers lavapipe in the registry,
which an elevated Vulkan loader reads.

+++ src/world/dem.cpp
@@ -258,7 +258,7 @@ float Dem::stored_sample(const Tile& t, DemCell cell, std::int64_t row,
 double Dem::sample(DemCell cell, std::int64_t row, std::int64_t column, int depth) {
     const Tile& t = tile(cell);
     if (row < t.rows && column < t.columns) {
-        return stored_sample(t, cell, row, column);
+        return static_cast<double>(stored_sample(t, cell, row, column));
     }
     // Past the south or east edge: the sample is the next tile's.
     const std::int64_t latitude =
